price_prediction/pyproject.toml
[project]
name = "price-prediction"
version = "0.1.0"
description = "NZ electricity price prediction system for paper mill load optimization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "autogluon.timeseries>=1.0",
    "requests>=2.32.5",
    "httpx>=0.28.1",
    "holidays>=0.92",
    "polars>=1.38.1",
    "rich>=14.3.3",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "tabulate>=0.10.0",
    "click>=8.0",
    "plotext>=5.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
